+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# [3.22.0](https://github.com/quintype/quintype-node-framework/compare/v3.20.1...v3.22.0) (2020-02-28)
+
+
+### Features
+
+* **akamai:** Add cache headers for akamai ([#108](https://github.com/quintype/quintype-node-framework/issues/108)) ([858bd47](https://github.com/quintype/quintype-node-framework/commit/858bd47))
+* Add cacheTag for story template, tag and section. ([#105](https://github.com/quintype/quintype-node-framework/issues/105)) ([af66b13](https://github.com/quintype/quintype-node-framework/commit/af66b13))
+
+
+
 # [3.21.0](https://github.com/quintype/quintype-node-framework/compare/v3.20.1...v3.21.0) (2020-02-27)